I cleaned up the old project and redesigned the views
