+++ app.js
@@ -7,8 +7,6 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
-var retreatantsRouter = require('./routes/retreatants');
 
 var app = express();
 
@@ -23,8 +21,6 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
-app.use('/users', usersRouter);
-app.use('/retreatants', retreatantsRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
