+++ views/error.pug
@@ -2,7 +2,7 @@ extends layout
 
 block content
   div(class='container')
-    div(class='row' )
+    div(class='col-md-9' )
       h2= message
       h3= error.status
       pre #{error.stack}
