+++ views/header.pug
@@ -1,21 +1,20 @@
 
 nav(class="navbar navbar-expand navbar-dark bg-dark")
-  a(class="navbar-brand" href="#") PDD Import
-  button(class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExample02" aria-controls="navb" aria-expanded="false" aria-label="Toggle navigation")
-    span(class="navbar-toggler-icon")
-  
-
-  div(class="collapse navbar-collapse" id="navb")
-    ul(class="navbar-nav mr-auto")
-      li(class="nav-item active")
-        a(class="nav-link" href="/") Home 
-          span(class="sr-only">(current))
+  div.container
+    a(class="navbar-brand" href="#")  Inscription aux Retraites - Import BD
+    button(class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExample02" aria-controls="navb" aria-expanded="false" aria-label="Toggle navigation")
+      span(class="navbar-toggler-icon")
     
-      li(class="nav-item")
-        if isSignedIn
-          a(class="nav-link" href="/users/signout") Logout
-        else
-          a(class="nav-link" href="/users") Login
+
+    div(class="collapse navbar-collapse" id="navb")
+      ul(class="navbar-nav mr-auto")
+        li(class="nav-item active")
+          a(class="nav-link" href="/") Home 
+            span(class="sr-only">(current))
+      
+        li(class="nav-item")
+          if isSignedIn
+            a(class="nav-link" href="/logout") Logout
   br
   br
