+++ views/instructions.pug
@@ -1,0 +1,8 @@
+div.jumbotron
+  h1.display-4 Welcome
+  p.lead Cette page permet de charger le fichier liste des inscrits aux retraites.
+  <hr class="my-4">
+  h3 Instructions
+  p Se logguer et charger la liste des inscrits. Seuls les fichiers json sont acceptes
+  img(src="images/example.png" alt="example de fichier accepte" class='w-50')
+
